+++ src/main/java/demo/codeexample/project/infrastructure/adapters/in/ProjectController.java
@@ -2,7 +2,7 @@ package demo.codeexample.project.infrastructure.adapters.in;
 
 import demo.codeexample.project.CreateProjectDto;
 import demo.codeexample.project.application.in.ProjectUseCase;
-import demo.codeexample.security.CurrentUserService;
+import demo.codeexample.auth.application.CurrentUserService;
 import demo.codeexample.user.UserLookup;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
